fixed StakeExpenditureDialog console errors
