+++ src/modules/dashboard/components/Dialogs/StakeExpenditureDialog/index.ts
@@ -1,1 +1,2 @@
 export { default } from './StakeExpenditureDialog';
+export { default as StakeExpenditureDialogForm } from './StakeExpenditureDialogForm';
