+++ src/modules/dashboard/components/ExpenditurePage/ExpenditureSettings/ExpenditureSettings.tsx
@@ -9,14 +9,16 @@ import {
   FormSection,
 } from '~core/Fields';
 import { Colony, useLoggedInUser } from '~data/index';
-import styles from './ExpenditureSettings.css';
 import UserAvatar from '~core/UserAvatar';
 import { tokens as tokensData } from './constants';
 import { COLONY_TOTAL_BALANCE_DOMAIN_ID } from '~constants';
 import UserMention from '~core/UserMention';
 import ColorTag, { Color } from '~core/ColorTag';
 import DomainDropdown from '~core/DomainDropdown';
+import { ExpenditureTypes } from '~pages/ExpenditurePage/types';
+
 import BalanceSelect from './BalanceSelect';
+import styles from './ExpenditureSettings.css';
 
 export const MSG = defineMessages({
   typeLabel: {
@@ -48,11 +50,11 @@ export const MSG = defineMessages({
 const expeditureTypes = [
   {
     label: MSG.optionAdvanced,
-    value: 'advanced',
+    value: ExpenditureTypes.Advanced,
   },
   {
     label: MSG.split,
-    value: 'split',
+    value: ExpenditureTypes.Split,
   },
 ];
