+++ src/modules/pages/components/ExpenditurePage/types.ts
@@ -1,0 +1,4 @@
+export enum ExpenditureTypes {
+  Advanced = 'advanced',
+  Split = 'split',
+}
